refactor: move static Brush methods into own Brushes class
