+++ src/vme/core/core/brushes/brush_loader.h
@@ -1,9 +1,9 @@
 #pragma once
 
-#include <nlohmann/json.hpp>
 #include <array>
 #include <filesystem>
 #include <memory>
+#include <nlohmann/json.hpp>
 #include <optional>
 #include <stack>
 #include <string>
@@ -37,28 +37,27 @@ class BrushLoader
     void parseBorderRules(BorderBrush &brush, const nlohmann::json &ruleJson);
     std::unique_ptr<BorderRuleAction> parseBorderRuleAction(const nlohmann::json &actionJson);
 
-    BorderBrush parseBorderBrush(const nlohmann::json &borderJson);
-    WallBrush parseWallBrush(const nlohmann::json &wallJson);
-    std::optional<DoodadBrush> parseDoodadBrush(const nlohmann::json &doodadJson);
-    std::optional<MountainBrush> parseMountainBrush(const nlohmann::json &mountainJson);
+    BorderBrush parseBorderBrush(const nlohmann::json &brush);
+    WallBrush parseWallBrush(const nlohmann::json &brush);
+    std::optional<DoodadBrush> parseDoodadBrush(const nlohmann::json &brush);
+    std::optional<MountainBrush> parseMountainBrush(const nlohmann::json &brush);
 
     void logError(std::string message);
 
-    static std::array<uint32_t, 12> parseBorderIds(const nlohmann::json &borderJson);
+    // 12 = BORDER_COUNT_FOR_GROUND_TILE. We don't use it explicitly because then we would need to include border_brush.h here.
+    static std::array<uint32_t, 12> parseBorderIds(const nlohmann::json &borderJson); // NOLINT
     static vme_unordered_map<uint32_t, BorderType> parseExtraBorderIds(const nlohmann::json &extrasJson);
 
     std::stack<std::string> stackTrace;
 };
 
 template <typename T>
-inline T jsonGetOrDefault(const nlohmann::json &j, std::string key, T defaultValue)
+inline T jsonGetOrDefault(const nlohmann::json &j, const std::string &key, T defaultValue)
 {
     if (j.contains(key))
     {
         return j.at(key).get<T>();
     }
-    else
-    {
-        return defaultValue;
-    }
+
+    return defaultValue;
 }
